修改id找不到的bug

+++ common/models/Order.php
@@ -289,7 +289,7 @@ class Order extends \yii\db\ActiveRecord
             if (!is_null($snapshot)) {
                 //查找当前的snapshot是否还有订单在引用
                 $isQuoteExist = self::find()->select('id,combo_id')->where(['combo_id' => $snapshot->id])->asArray()->all();
-                if (count($isQuoteExist) <= 1 && $isQuoteExist['id'] == $snapshot->id) {
+                if (count($isQuoteExist) <= 1 && isset($isQuoteExist['id']) && $isQuoteExist['id'] == $snapshot->id) {
                     if ($isQuoteExist['snap_combo_id'] == 0 || is_null(Combo::findOne(['combo_id' => $isQuoteExist['snap_combo_id']]))) {
                         $snapshot->delete();
                     }
